Completed user_select_difficulty
Completed user_select_difficulty(), this method is created to manually
check the input of the user via string comparison. This is helpful to
set the amount of white space the user is willing to play with. The
function works by comparing string input with the preset difficulty
names such as Easy, Medium, Hard. This method also catches errors and
such.